sdsds

+++ moviews/model.go
@@ -3,11 +3,11 @@ package moviews
 type Movie struct {
 	Id     int64   `json:"id"`
 	Name   string  `json:"name"`
-	Rating float64 `json:"rating"`
+	Rating int64 `json:"rating"`
 }
 
 type MovieUpdate struct {
 	Id     int64   `json:"id"`
 	Name   *string  `json:"name"`
-	Rating *float64 `json:"rating"`
+	Rating *int64 `json:"rating"`
 }
